add a simple set data structure
We work with a few sets, so add a data structure for it to avoid
ad-hoc implementations in other modules.  This is simplify a list
as the sets we're working with tends to be quite small: typically
less than 10.  As long as we don't need to support more, the perf
should be good enough.